requirements.txt
kaggle==1.6.6
pandas==2.2.2
dbt-core==1.8.1
dbt-postgres==1.8.1
argparse==1.4.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1